SDT-223: Remove UnsupportedOperationException from BulkFeedbackEnricher
The BulkFeedbackEnricher was intended to throw an UnsupportedOperationException if the response returned for an individual request contained a non-rejected status and no response from MCOL (i.e. the responseDetail tag is empty).  In practise the exception was never thrown because the regular expression used to detect such an occurrence expects the requestId attribute to be the last attribute in the response tag and this was never the case.  Updating the libraries used by SDT has changed the attribute order, so exceptions are being thrown when previously they weren't.  Removed the exception to ensure the same behaviour as before.